+++ test/test_roscpp/test/launch/missing_call_to_shutdown.xml
@@ -1,4 +1,3 @@
 <launch>
-  <test test-name="missing_call_to_shutdown_init_only" pkg="test_roscpp" type="test_roscpp-missing_call_to_shutdown" args="0"/>
-  <test test-name="missing_call_to_shutdown_init_and_start" pkg="test_roscpp" type="test_roscpp-missing_call_to_shutdown" args="1"/>
+  <test test-name="missing_call_to_shutdown" pkg="test_roscpp" type="test_roscpp-missing_call_to_shutdown"/>
 </launch>
