+++ test/test_roscpp/test/src/missing_call_to_shutdown.cpp
@@ -29,112 +29,24 @@
 
 /* Author: David Gossow */
 
-/*
- * Call ros::start() explicitly, but never call ros::shutdown().
- * ros::shutdown should be called automatically in this case.
- */
-
+#include <gtest/gtest.h>
 #include <cstdlib>
-#include <ros/ros.h>
 
-namespace ros
+TEST(roscpp, missingCallToShutdownInitOnly)
 {
-namespace console
-{
-extern bool g_shutting_down;
-}
+  int exit_code = system("rosrun test_roscpp test_roscpp-missing_call_to_shutdown_impl 0");
+  EXPECT_EQ(0, exit_code);
 }
 
-namespace
+TEST(roscpp, missingCallToShutdownInitAndStart)
 {
-
-enum TestId
-{
-  InitOnly = 0,
-  InitAndStart = 1
-};
-
-TestId test_id = InitOnly;
-
-void atexitCallback()
-{
-  bool hasError = false;
-
-  if (ros::ok())
-  {
-    std::cerr << "ERROR: ros::ok() returned true!" << std::endl;
-    hasError = true;
-  }
-  if (!ros::isShuttingDown())
-  {
-    std::cerr << "ERROR: ros::isShuttingDown() returned false!" << std::endl;
-    hasError = true;
-  }
-  if (ros::isStarted())
-  {
-    std::cerr << "ERROR: ros::isStarted() returned true!" << std::endl;
-    hasError = true;
-  }
-
-  if (!ros::isInitialized())
-  {
-    std::cerr << "ERROR: ros::isInitialized() returned false, although ros::init was called!" << std::endl;
-    std::cerr << "Due to legacy reasons, it should return true, even after ROS has been de-initialized." << std::endl;
-    hasError = true;
-  }
-
-  if (!ros::console::g_shutting_down)
-  {
-    std::cerr << "ERROR: ros::console::g_shutting_down returned false, but it should have been automatically shut down." << std::endl;
-    hasError = true;
-  }
-
-  if (hasError)
-  {
-    std::_Exit(1);
-  }
-}
+  int exit_code = system("rosrun test_roscpp test_roscpp-missing_call_to_shutdown_impl 1");
+  EXPECT_EQ(0, exit_code);
 }
 
 int
 main(int argc, char** argv)
 {
-  if ( argc > 1 )
-  {
-    test_id = static_cast<TestId>(atoi(argv[1]));
-  }
-
-  // Register atexit callbak which will be executed after ROS has been de-initialized.
-  if (atexit(atexitCallback) != 0)
-  {
-    std::cerr << "Failed to register atexit callback." << std::endl;
-    return 1;
-  }
-
-  switch (test_id)
-  {
-    case InitOnly:
-      // Test case 0: Call ros::init() explicitly, but never call ros::shutdown().
-      // ros::deInit should be called automatically in this case.
-      ros::init(argc, argv, "missing_call_to_shutdown" );
-      break;
-    case InitAndStart:
-      // Test case 1: Call ros::init() and ros::start() explicitly, but never call ros::shutdown().
-      // ros::shutdown should be called automatically in this case.
-      ros::init(argc, argv, "missing_call_to_shutdown" );
-      ros::start();
-      break;
-    default:
-      std::cerr << "Invalid test id: " << test_id << std::endl;
-      return 1;
-      break;
-  }
-
-  if (!ros::ok())
-  {
-    std::cerr << "Failed to start ROS." << std::endl;
-    return 1;
-  }
-  
-  return 0;
+  testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
 }
